requirements.txt
torch>=1.5.0
torchvision>=0.6.0

Pillow>=7.0.0
scikit-image>=0.16.2
matplotlib>=3.1.2

click>=7.0
